requirements.txt
Django==4.2.21
djangorestframework==3.14.0
Pillow==10.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-cors-headers==4.3.1
django-filter==23.5
djangorestframework-simplejwt==5.3.1
django-storages==1.14.2
django-cleanup==8.0.0
django-widget-tweaks==1.5.0
django-bootstrap5==23.4
django-crispy-forms==2.1
crispy-bootstrap5==2023.10
django-allauth==0.61.1
django-environ==0.11.2
django-debug-toolbar==4.3.0
django-extensions==3.2.3
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.6
redis==5.0.1
whitenoise==6.6.0
gunicorn==21.2.0
boto3==1.34.34 